esp_idf: add shared harness and use it from the http_client pytest
Move the modular part of the conftest.py found in http_client to an
optional harness pare part of the pytest-pouch plugin.
